Move simple tests into separate spec
These tests are simple enough to be tested usig an exact workflow expectation
